Vector to matrix implementation and changes in the tests accordingly

+++ src/engine/mod.rs
@@ -7,6 +7,6 @@ pub type EngineResult<T> = Result<T, EngineError>;
 
 pub trait EngineStrategy<T> {
     fn get_subject(&self, index: usize) -> EngineResult<T>;
-    fn get_predicate(&self, index: usize) -> EngineResult<T>;
+    fn get_predicate(&self, index: u8) -> EngineResult<T>;
     fn get_object(&self, index: usize) -> EngineResult<T>;
 }
